replace Shop with self inside shop model
following the README
when ShopifySessionRepository.storage = Shop is enabled in
config/initializers/shopify_session_repository.rb
and the Shop model is modified
it gives error :
ArgumentError (A copy of Shop has been removed from the module tree but is still active!)
in development.

to solve the error Shop is replaced with identifier self.
this solves the error in a simple manner

so changes are made to the gerenrated file
config/initializers/shopify_session_repository.rb