Updated README to include changes to @shop_session
the @shop_session variable is now referred to as @current_shopify_session in v13.0.0

Comparing the definition of this variable in v13.0.0 (line 29):
https://github.com/Shopify/shopify_app/blob/v13.0.0/lib/shopify_app/controller_concerns/login_protection.rb

to v12.0.7 (line 29):
https://github.com/Shopify/shopify_app/blob/v12.0.7/lib/shopify_app/controller_concerns/login_protection.rb

There is a change that occurs. We set the value of @current_shopify_session to @shop_session if it exists, although I've experienced that @shop_session evaluates to nil if referenced in my controllers, while @current_shopify_session evaluates to what is expected.